+++ lib/paper_ripple.dart
@@ -1,7 +1,7 @@
 // DO NOT EDIT: auto-generated with `pub run custom_element_apigen:update`
 
 /// Dart API for the polymer element `paper_ripple`.
-@HtmlImport('package:paper_elements/paper_ripple_nodart.html')
+@HtmlImport('paper_ripple_nodart.html')
 library paper_elements.paper_ripple;
 
 import 'dart:html';