+++ lib/paper_toggle_button.dart
@@ -1,7 +1,7 @@
 // DO NOT EDIT: auto-generated with `pub run custom_element_apigen:update`
 
 /// Dart API for the polymer element `paper_toggle_button`.
-@HtmlImport('package:paper_elements/paper_toggle_button_nodart.html')
+@HtmlImport('paper_toggle_button_nodart.html')
 library paper_elements.paper_toggle_button;
 
 import 'dart:html';
@@ -10,6 +10,7 @@ import 'package:web_components/custom_element_proxy.dart';
 import 'package:web_components/html_import_annotation.dart';
 import 'package:custom_element_apigen/src/common.dart' show PolymerProxyMixin, DomProxyMixin;
 import 'paper_radio_button.dart';
+import 'package:core_elements/core_a11y_keys.dart';
 
 /// `paper-toggle-button` provides a ON/OFF switch that user can toggle the state
 /// by tapping or by dragging the swtich.
@@ -20,40 +21,38 @@ import 'paper_radio_button.dart';
 ///
 /// Styling toggle button:
 ///
-/// To change the ink color for checked state:
+/// To change the toggle color:
 ///
-///     paper-toggle-button::shadow paper-radio-button::shadow #ink[checked] {
+///     paper-toggle-button::shadow .toggle {
+///       background-color: #9c27b0;
+///     }
+///
+/// To change the ink color:
+///
+///     paper-toggle-button::shadow .toggle-ink {
+///       color: #009688;
+///     }
+///
+/// To change the checked toggle color:
+///
+///     paper-toggle-button::shadow [checked] .toggle {
+///       background-color: #4285f4;
+///     }
+///
+/// To change the checked ink color:
+///
+///     paper-toggle-button::shadow [checked] .toggle-ink {
 ///       color: #4285f4;
 ///     }
 ///
-/// To change the radio checked color:
+/// To change the toggle bar and toggle button colors separately:
 ///
-///     paper-toggle-button::shadow paper-radio-button::shadow #onRadio {
-///       background-color: #4285f4;
+///     paper-toggle-button::shadow .toggle-bar {
+///       background-color: #5677fc;
 ///     }
 ///
-/// To change the bar color for checked state:
-///
-///     paper-toggle-button::shadow #toggleBar[checked] {
-///       background-color: #4285f4;
-///     }
-///
-/// To change the ink color for unchecked state:
-///
-///     paper-toggle-button::shadow paper-radio-button::shadow #ink {
-///       color: #b5b5b5;
-///     }
-///
-/// To change the radio unchecked color:
-///
-///     paper-toggle-button::shadow paper-radio-button::shadow #offRadio {
-///       border-color: #b5b5b5;
-///     }
-///
-/// To change the bar color for unchecked state:
-///
-///     paper-toggle-button::shadow #toggleBar {
-///       background-color: red;
+///     paper-toggle-button::shadow .toggle-button {
+///       background-color: #9c27b0;
 ///     }
 @CustomElementProxy('paper-toggle-button')
 class PaperToggleButton extends HtmlElement with DomProxyMixin, PolymerProxyMixin {